refactor(engine): remove old layout to replace by restruct layout
